Organize files, make update email nested
