+++ webhook/tests/test_query_resulsts.py
@@ -1,19 +1,10 @@
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
-from google.cloud import ndb
 from main import app
-from models import UserEmail
-
-client = TestClient(app)
-
-
-def test_index():
-    r = client.get('/')
-    assert r.status_code == 200
-    assert r.json() == {"data": "Hello World"}
 
 
 correct_headers = {"X-Appengine-Inbound-Appid": "virustotal-step-2020"}
+client = TestClient(app)
 
 
 def test_query_results_empty_list():
@@ -141,54 +132,3 @@ def test_query_results_missing_field_in_data():
             })
 
         assert r.status_code == 422
-
-
-def test_email_address_new(monkeypatch):
-    data = {
-        "api_key": "abcabc",
-        "email": "some@email.example"
-    }
-
-    def put(self):
-        assert self == UserEmail(api_key=ndb.Key(
-            "UserEmail", data["api_key"]), email=data["email"])
-
-    monkeypatch.setattr(UserEmail, 'put', put)
-
-    def get(self):
-        return None
-
-    monkeypatch.setattr(ndb.Key, 'get', get)
-
-    r = client.post(
-        '/email-address/',
-        json=data
-    )
-    print(r.text)
-    assert r.status_code == 201
-
-
-def test_email_address_update(monkeypatch):
-    data = {
-        "api_key": "abcabc",
-        "email": "some@email.example"
-    }
-
-    def put(self):
-        assert self == UserEmail(api_key=ndb.Key("UserEmail", data["api_key"]),
-                                 email=data["email"])
-
-    monkeypatch.setattr(UserEmail, 'put', put)
-
-    def get(self):
-        return UserEmail(api_key=ndb.Key("UserEmail", data["api_key"]),
-                         email="random@email.example")
-
-    monkeypatch.setattr(ndb.Key, 'get', get)
-
-    r = client.post(
-        '/email-address/',
-        json=data
-    )
-    print(r.text)
-    assert r.status_code == 200
